requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
lxml==5.3.1
multipledispatch==1.0.0
numpy==2.1.1
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32-ctypes==0.2.2
reportlab==4.2.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
XlsxWriter==3.2.2
